requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
eventlet==0.36.1
python-socketio==5.9.0
openai==1.12.0
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0

# Base de datos (para Azure/AWS PostgreSQL)
psycopg2-binary>=2.9.10
SQLAlchemy==2.0.23

# Servidor de producción (para Azure App Service)
gunicorn==21.2.0

# Cloud SDKs (opcional)
# boto3==1.34.14  # Para AWS
# azure-identity==1.15.0  # Para Azure

# Redis (para Socket.IO multi-servidor - futuro)
redis==5.0.1

